Created pagination comp for pagination
