+++ package.json
@@ -27,7 +27,6 @@
     "react-hook-form": "^7.27.0",
     "react-icons": "^4.3.1",
     "react-image-file-resizer": "^0.4.7",
-    "react-infinite-scroll-component": "^6.1.0",
     "react-infinite-scroller": "^1.2.5",
     "react-quill": "^0.0.2",
     "react-render-html": "^0.6.0",
