+++ pages/api/data.js
@@ -1,3 +1,5 @@
+import { getSession } from 'next-auth/react';
+
 import prisma from '../../lib/prisma';
 
 export default async (req, res) => {
@@ -5,21 +7,43 @@ export default async (req, res) => {
 
   switch (method) {
     case 'GET':
-      const { skip } = query;
-      let len = await prisma.links.findMany({});
+      const { skip, user } = query;
+      const session = await getSession({ req });
 
-      await prisma.links
-        .findMany({
-          skip: parseInt(skip),
-          take: 2,
-        })
-        .then(async (re) => {
-          res.json({
-            status: 200,
-            data: re,
-            len: len.length,
+      if (!user) {
+        let len = await prisma.links.findMany({});
+
+        await prisma.links
+          .findMany({
+            skip: parseInt(skip),
+            take: 2,
+          })
+          .then(async (re) => {
+            res.json({
+              status: 200,
+              data: re,
+              len: len.length,
+            });
           });
+      } else {
+        let len = await prisma.links.findMany({
+          where: { userName: session?.user.name },
         });
+
+        await prisma.links
+          .findMany({
+            where: { userName: session?.user.name },
+            skip: parseInt(skip),
+            take: 2,
+          })
+          .then(async (re) => {
+            res.json({
+              status: 200,
+              data: re,
+              len: len.length,
+            });
+          });
+      }
       break;
     default:
       res.status(400).json({
